Spec based proposal
This aims to make a spec based proposal that also addresses the
feedbak given to prior proposals and to the specifics of the repolinter
that were not objected to by anyone else.

I also updated the repolinter to make it run against all projects.  Some
hackery with copying the repolinter into the local cloned repo was
required as repolinter does not allow pointing to an external config.

Required content is not automatically checked, but is part of the spec.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>