Gets the list of PRs as per the configuration
The weekly developer newsletter lists PRs from across
Hyperledger and Hyperledger Labs repositories which are
noteworthy for that week. This lab tool helps in fetching
all the PRs from Hyperledger and Hyperledger Labs that were
raised in last 7 days (closed or not).

The tool can be extended later to have intelligent filtering.
But to start with, the tool prints out all the PRs in a
html page. This should make at least the manual look through
easier than before.

Note that the tool may expect GITHUB_TOKEN for making
frequent requests. Github would otherwise block
the requests if it's too frequent.

Signed-off-by: Vyshnavi Bendi <bendi.vyshnavi@walmartlabs.com>